pytorch 0.4.0

+++ torch_cluster/utils/perm.py
@@ -3,7 +3,7 @@ import torch
 
 def randperm(row, col):
     # Randomly reorder row and column indices.
-    edge_rid = torch.randperm(row.size(0)).type_as(row)
+    edge_rid = torch.randperm(row.size(0))
     return row[edge_rid], col[edge_rid]
 
 
@@ -16,7 +16,7 @@ def sort_row(row, col):
 
 def randperm_sort_row(row, col, num_nodes):
     # Randomly change row indices to new values.
-    node_rid = torch.randperm(num_nodes).type_as(row)
+    node_rid = torch.randperm(num_nodes)
     row = node_rid[row]
 
     # Sort row and column indices row-wise.
